app/requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
psycopg2-binary
setuptools>=78.1.1